[JT-80] test: test code 수정
