+++ module-application/app-api/src/main/java/shop/jtoon/payment/application/PaymentService.java
@@ -1,11 +1,8 @@
 package shop.jtoon.payment.application;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.math.BigDecimal;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
 import shop.jtoon.dto.MemberDto;
 import shop.jtoon.entity.Member;
 import shop.jtoon.member.application.MemberService;
@@ -15,41 +12,44 @@ import shop.jtoon.payment.request.PaymentReq;
 import shop.jtoon.payment.response.PaymentRes;
 import shop.jtoon.service.IamportService;
 
+import java.math.BigDecimal;
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class PaymentService {
 
-	private final IamportService iamportService;
-	private final PaymentInfoService paymentInfoService;
-	private final MemberCookieService memberCookieService;
-	private final MemberService memberService;
+    private final IamportService iamportService;
+    private final PaymentInfoService paymentInfoService;
+    private final MemberCookieService memberCookieService;
+    private final MemberService memberService;
 
-	@Transactional
-	public BigDecimal validateAndCreatePayment(PaymentReq paymentReq, MemberDto memberDto) {
-		Member member = memberService.findById(memberDto.id());
-		iamportService.validateIamport(paymentReq.impUid(), paymentReq.amount());
-		paymentInfoService.validatePaymentInfo(paymentReq);
-		paymentInfoService.createPaymentInfo(paymentReq, member);
-		memberCookieService.createMemberCookie(paymentReq.itemName(), member);
+    @Transactional
+    public BigDecimal validateAndCreatePayment(PaymentReq paymentReq, MemberDto memberDto) {
+        Member member = memberService.findByEmail(memberDto.email());
+        iamportService.validateIamport(paymentReq.impUid(), paymentReq.amount());
+        paymentInfoService.validatePaymentInfo(paymentReq);
+        paymentInfoService.createPaymentInfo(paymentReq, member);
+        memberCookieService.createMemberCookie(paymentReq.itemName(), member);
 
-		return paymentReq.amount();
-	}
+        return paymentReq.amount();
+    }
 
-	@Transactional
-	public void cancelPayment(CancelReq cancelReq) {
-		iamportService.validateIamport(cancelReq.impUid(), cancelReq.checksum());
-		iamportService.cancelIamport(
-			cancelReq.impUid(),
-			cancelReq.reason(),
-			cancelReq.checksum(),
-			cancelReq.refundHolder()
-		);
-	}
+    @Transactional
+    public void cancelPayment(CancelReq cancelReq) {
+        iamportService.validateIamport(cancelReq.impUid(), cancelReq.checksum());
+        iamportService.cancelIamport(
+                cancelReq.impUid(),
+                cancelReq.reason(),
+                cancelReq.checksum(),
+                cancelReq.refundHolder()
+        );
+    }
 
-	public List<PaymentRes> getPayments(ConditionReq conditionReq, MemberDto memberDto) {
-		Member member = memberService.findById(memberDto.id());
+    public List<PaymentRes> getPayments(ConditionReq conditionReq, MemberDto memberDto) {
+        Member member = memberService.findByEmail(memberDto.email());
 
-		return paymentInfoService.getPaymentsInfo(conditionReq.merchantsUid(), member);
-	}
+        return paymentInfoService.getPaymentsInfo(conditionReq.merchantsUid(), member);
+    }
 }
