+++ module-application/app-api/src/test/java/shop/jtoon/factory/PaymentSnippetFactory.java
@@ -1,4 +1,4 @@
-package shop.jtoon.payment.factory;
+package shop.jtoon.factory;
 
 import org.springframework.restdocs.payload.RequestFieldsSnippet;
