+++ module-application/app-api/src/test/java/shop/jtoon/payment/presentation/PaymentControllerTest.java
@@ -10,20 +10,19 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
+import shop.jtoon.config.WithCurrentUser;
 import shop.jtoon.entity.Member;
 import shop.jtoon.entity.PaymentInfo;
 import shop.jtoon.exception.IamportException;
-import shop.jtoon.payment.factory.PaymentFactory;
-import shop.jtoon.payment.factory.PaymentSnippetFactory;
+import shop.jtoon.factory.MemberFactory;
+import shop.jtoon.factory.PaymentFactory;
+import shop.jtoon.factory.PaymentSnippetFactory;
 import shop.jtoon.payment.request.PaymentReq;
 import shop.jtoon.repository.MemberRepository;
 import shop.jtoon.repository.PaymentInfoRepository;
-import shop.jtoon.security.service.AuthenticationService;
 import shop.jtoon.service.IamportService;
 
 import java.math.BigDecimal;
@@ -57,9 +56,6 @@ class PaymentControllerTest {
     @Autowired
     private MemberRepository memberRepository;
 
-    @Autowired
-    private AuthenticationService authenticationService;
-
     @MockBean
     private IamportService iamportService;
 
@@ -67,18 +63,16 @@ class PaymentControllerTest {
     private String merchantUid;
     private Member member;
 
-
     @BeforeEach
     void beforeEach() {
         impUid = "impUid123";
         merchantUid = "merchant123";
-        member = PaymentFactory.createMember("example123@naver.com");
+        member = MemberFactory.createMember();
         memberRepository.save(member);
-        Authentication auth = authenticationService.getAuthentication(member.getEmail());
-        SecurityContextHolder.getContext().setAuthentication(auth);
     }
 
     @DisplayName("POST: /payments/validation - 결제 승인 후 결제 정보에 대해 검증 및 생성이 성공적으로 됐을 때, - Amount")
+    @WithCurrentUser
     @Test
     void validatePayment_Amount() throws Exception {
         // Given
@@ -101,6 +95,7 @@ class PaymentControllerTest {
     }
 
     @DisplayName("POST: /payments/validation - 결제 승인 정보가 조회되지 않거나 실제 결제 금액과 요청 금액이 다를 때, - IamportException")
+    @WithCurrentUser
     @Test
     void validatePayment_IamportException() throws Exception {
         // Given
@@ -122,6 +117,7 @@ class PaymentControllerTest {
     }
 
     @DisplayName("POST: /payments/validation - 결제 정보의 쿠키 가격과 실제 서버에서 알고 있는 쿠키 가격이 다를 때, - InvalidRequestException")
+    @WithCurrentUser
     @Test
     void validatePayment_InvalidRequestException() throws Exception {
         // Given
@@ -144,6 +140,7 @@ class PaymentControllerTest {
     }
 
     @DisplayName("POST: /payments/validation - 포트원 결제 고유번호가 중복 됐을 때, - ImpUid DuplicatedException")
+    @WithCurrentUser
     @Test
     void validatePayment_ImpUid_DuplicatedException() throws Exception {
         // Given
@@ -168,6 +165,7 @@ class PaymentControllerTest {
     }
 
     @DisplayName("POST: /payments/validation - 가맹점 주문번호가 중복 됐을 때, - MerchantUid DuplicatedException")
+    @WithCurrentUser
     @Test
     void validatePayment_MerchantUid_DuplicatedException() throws Exception {
         // Given
